Mainproject/requirements.txt
beautifulsoup4
requests
playwright
reportlab
Flask
Pillow


